+++ be1-go/internal/popserver/handler/answer_test.go
@@ -1,11 +1,10 @@
-package message
+package handler
 
 import (
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
 	"github.com/stretchr/testify/require"
-	"go.dedis.ch/kyber/v3"
 	"go.dedis.ch/kyber/v3/sign/schnorr"
 	"popstellar/crypto"
 	"popstellar/internal/popserver/database"
@@ -105,23 +104,4 @@ func Test_handleMessagesByChannel(t *testing.T) {
 		})
 	}
 
-}
-
-type Keypair struct {
-	Public     kyber.Point
-	PublicBuf  []byte
-	Private    kyber.Scalar
-	PrivateBuf []byte
-}
-
-func GenerateKeyPair(t *testing.T) Keypair {
-	secret := crypto.Suite.Scalar().Pick(crypto.Suite.RandomStream())
-	point := crypto.Suite.Point().Mul(secret, nil)
-
-	publicBuf, err := point.MarshalBinary()
-	require.NoError(t, err)
-	privateBuf, err := secret.MarshalBinary()
-	require.NoError(t, err)
-
-	return Keypair{point, publicBuf, secret, privateBuf}
 }
