+++ be1-go/internal/popserver/handler/coin.go
@@ -1,4 +1,4 @@
-package channel
+package handler
 
 import (
 	"popstellar/internal/popserver/database"
