+++ be1-go/internal/popserver/handler/election.go
@@ -1,4 +1,4 @@
-package channel
+package handler
 
 import (
 	"bytes"
