+++ be1-go/internal/popserver/handler/root.go
@@ -1,4 +1,4 @@
-package channel
+package handler
 
 import (
 	"encoding/base64"
