+++ be1-go/internal/popserver/utils/utils.go
@@ -1,4 +1,4 @@
-package util
+package utils
 
 import (
 	"github.com/rs/zerolog"
